+++ flake.nix
@@ -7,8 +7,6 @@
     nixpkgs.url = "github:nixos/nixpkgs/nixpkgs-unstable";
     stable.url = "github:nixos/nixpkgs/release-24.05";
 
-    flaresolverr-pin.url = "github:nixos/nixpkgs/ebbc0409688869938bbcf630da1c1c13744d2a7b";
-
     # Secrets Management <3
     agenix = {
       inputs = {
@@ -246,7 +244,7 @@
         flake-utils.follows = "flake-utils";
         nixpkgs.follows = "nixpkgs";
       };
-      url = "github:NuschtOS/nixos-modules/70d3b88b52af86098b7e008622871265ec39c800";
+      url = "github:NuschtOS/nixos-modules";
     };
 
     sbomnix = {