+++ modules/flaresolverr/default.nix
@@ -1,13 +1,8 @@
-{ pkgs, self, ... }:
-let
-  inherit (pkgs) system;
-  inherit (self.inputs.flaresolverr-pin.legacyPackages.${system}) flaresolverr;
-in
+{ self, ... }:
 {
   services.flaresolverr = {
     enable = true;
     openFirewall = true;
-    package = flaresolverr;
     inherit (self.common.config.services.flaresolverr) port;
   };
 }